+++ packages/docs/components/Fetch/stories/simple/app.js
@@ -1,0 +1,6 @@
+import { registerComponent } from '@studiometa/js-toolkit';
+import { Action, Fetch, Transition } from '@studiometa/ui';
+
+registerComponent(Action);
+registerComponent(Fetch);
+registerComponent(Transition);