+++ packages/tests/__utils__/h.ts
@@ -1,3 +1,5 @@
+const { isArray } = Array;
+
 /**
  * Functional DOM node creation.
  */
@@ -17,17 +19,24 @@ export function h<T extends keyof HTMLElementTagNameMap = 'div'>(
 ): HTMLElementTagNameMap[T] {
   const el = document.createElement(tag);
 
-  if (Array.isArray(attributes) && children === null) {
+  if ((Array.isArray(attributes) || typeof attributes === 'string') && children === null) {
     children = attributes;
     attributes = {};
   }
 
   for (const [name, value] of Object.entries(attributes)) {
-    el.setAttribute(name.replaceAll(/([a-z])([A-Z])/g, '$1-$2').toLowerCase(), typeof value === 'string' ? value : JSON.stringify(value));
+    el.setAttribute(
+      name.replaceAll(/([a-z])([A-Z])/g, '$1-$2').toLowerCase(),
+      typeof value === 'string' ? value : JSON.stringify(value),
+    );
   }
 
   if (children) {
-    el.append(...children);
+    if (isArray(children)) {
+      el.append(...children);
+    } else {
+      el.append(children);
+    }
   }
 
   return el;
@@ -37,7 +46,7 @@ export function h<T extends keyof HTMLElementTagNameMap = 'div'>(
  * Connect the given element to a document instance in order
  * to set its `isConnected` property to `true`.
  */
-export function connectElement<T extends Node>(element:T):T {
+export function connectElement<T extends Node>(element: T): T {
   const doc = new Document();
   doc.append(element);
   return element;