+++ packages/tests/vitest.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from 'vitest/config'
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
@@ -10,10 +10,9 @@ export default defineConfig({
     setupFiles: ['./tests/__utils__/happydom.ts'],
     coverage: {
       provider: 'v8',
-      include: ['ui/**']
+      include: ['ui/**/*.ts'],
+      exclude: ['**/tests/**/*.ts', '**/ui/**/index.ts'],
     },
-    exclude: [
-      '**/.symfony/vendor/**'
-    ]
+    exclude: ['**/.symfony/vendor/**'],
   },
-})
+});