+++ packages/ui/index.ts
@@ -7,6 +7,7 @@ export * from './Cursor/index.js';
 export * from './Data/index.js';
 export * from './decorators/index.js';
 export * from './Draggable/index.js';
+export * from './Fetch/index.js';
 export * from './Figure/index.js';
 export * from './FigureVideo/index.js';
 export * from './Frame/index.js';